Save progress on v2 re-write
