+++ .vscode/settings.json
@@ -1,7 +1,8 @@
 {
   "editor.formatOnSave": true,
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
+    "source.fixAll.eslint": true,
+    "source.organizeImports": false
   },
   "eslint.packageManager": "pnpm",
   "typescript.tsdk": "node_modules/typescript/lib",
