+++ packages/connectors/.gitignore
@@ -6,3 +6,4 @@ metaMask/**
 mock/**
 safe/**
 walletConnect/**
+walletConnectV1/**
