+++ packages/connectors/src/walletConnect.test.ts
@@ -1,109 +1,7 @@
-import { foundry, goerli, mainnet } from '@wagmi/core/chains'
-import { testChains } from '@wagmi/core/internal/test'
-import WalletConnectProvider from '@walletconnect/ethereum-provider'
-import { UniversalProvider } from '@walletconnect/universal-provider'
-import { rest } from 'msw'
-import { setupServer } from 'msw/node'
-import {
-  afterAll,
-  afterEach,
-  beforeAll,
-  describe,
-  expect,
-  it,
-  vi,
-} from 'vitest'
-
-import { WalletConnectConnector } from './walletConnect'
-
-const defaultChains = [mainnet, goerli]
-
-const handlers = [
-  rest.get('https://*.bridge.walletconnect.org', (_req, res, ctx) => {
-    return res(
-      ctx.status(200),
-      ctx.json({
-        topic: '222781e3-3fad-4184-acde-077796bf0d3d',
-        type: 'sub',
-        payload: '',
-        silent: true,
-      }),
-    )
-  }),
-]
-
-const server = setupServer(...handlers)
+import { describe, it } from 'vitest'
 
 describe('WalletConnectConnector', () => {
-  beforeAll(() => {
-    server.listen({
-      onUnhandledRequest: 'warn',
-    })
-
-    const matchMedia = vi.fn().mockImplementation((query) => {
-      return {
-        matches: false,
-        media: query,
-        onchange: null,
-        addListener: vi.fn(), // deprecated
-        removeListener: vi.fn(), // deprecated
-        addEventListener: vi.fn(),
-        removeEventListener: vi.fn(),
-        dispatchEvent: vi.fn(),
-      }
-    })
-    vi.stubGlobal('matchMedia', matchMedia)
-  })
-
-  afterEach(() => server.resetHandlers())
-
-  afterAll(() => server.close())
-
-  it('inits', () => {
-    const connector = new WalletConnectConnector({
-      chains: testChains,
-      options: {
-        rpc: {
-          [foundry.id]: foundry.rpcUrls.default.http[0]!,
-        },
-      },
-    })
-    expect(connector.name).toEqual('WalletConnect')
-  })
-
-  it('should use v1 by default', async () => {
-    const connector = new WalletConnectConnector({
-      chains: defaultChains,
-      options: {
-        rpc: {
-          [foundry.id]: foundry.rpcUrls.default.http[0]!,
-        },
-      },
-    })
-    expect(await connector.getProvider()).instanceOf(WalletConnectProvider)
-  })
-
-  it('should use v1 as configured by options', async () => {
-    const connector = new WalletConnectConnector({
-      chains: defaultChains,
-      options: {
-        rpc: {
-          [foundry.id]: foundry.rpcUrls.default.http[0]!,
-        },
-        version: '1',
-      },
-    })
-    expect(await connector.getProvider()).instanceOf(WalletConnectProvider)
-  })
-
-  it('should use v2 as configured by options', async () => {
-    const connector = new WalletConnectConnector({
-      chains: defaultChains,
-      options: {
-        projectId: process.env.VITE_WC_PROJECT_ID!,
-        version: '2',
-      },
-    })
-    expect(await connector.getProvider()).instanceOf(UniversalProvider)
+  it('needs tests', () => {
+    // TODO
   })
 })
