+++ packages/connectors/src/walletConnect.ts
@@ -6,57 +6,44 @@ import {
   getClient,
   normalizeChainId,
 } from '@wagmi/core'
-import type WalletConnectProvider from '@walletconnect/ethereum-provider'
-import type {
-  UniversalProviderOpts,
-  UniversalProvider as UniversalProvider_,
-} from '@walletconnect/universal-provider'
-import type { Web3Modal } from '@web3modal/standalone'
+import type EthereumProviderType from '@walletconnect/ethereum-provider'
 import { providers } from 'ethers'
 import { getAddress, hexValue } from 'ethers/lib/utils.js'
 
 import { Connector } from './base'
 
-// Shared config for WalletConnect v2
-const defaultV2Config = {
-  namespace: 'eip155',
-  methods: [
-    'eth_sendTransaction',
-    'eth_sendRawTransaction',
-    'eth_sign',
-    'eth_signTransaction',
-    'eth_signTypedData',
-    'eth_signTypedData_v3',
-    'eth_signTypedData_v4',
-    'personal_sign',
-    'wallet_switchEthereumChain',
-    'wallet_addEthereumChain',
-  ],
-  events: ['accountsChanged', 'chainChanged'],
-}
+// -- Types --------------------------------------------------------------------
+type RpcMethod =
+  | 'eth_sendTransaction'
+  | 'eth_sendRawTransaction'
+  | 'eth_sign'
+  | 'eth_signTransaction'
+  | 'eth_signTypedData'
+  | 'eth_signTypedData_v3'
+  | 'eth_signTypedData_v4'
+  | 'personal_sign'
+  | 'wallet_switchEthereumChain'
+  | 'wallet_addEthereumChain'
 
-type UniversalProvider = InstanceType<typeof UniversalProvider_>
+type RpcEvent = 'accountsChanged' | 'chainChanged'
 
 type WalletConnectOptions = {
-  /** When `true`, uses default WalletConnect QR Code modal */
   qrcode?: boolean
-} & (
-  | ({
-      version?: '1'
-    } & ConstructorParameters<typeof WalletConnectProvider>[0])
-  | ({
-      /**
-       * Project ID associated with [WalletConnect account](https://cloud.walletconnect.com)
-       */
-      projectId: NonNullable<UniversalProviderOpts['projectId']>
-      version: '2'
-    } & Omit<UniversalProviderOpts, 'projectId'>)
-)
+  methods?: RpcMethod[]
+  events?: RpcEvent[]
+  projectId: string
+}
 
 type WalletConnectSigner = providers.JsonRpcSigner
 
+type ConnectorConfig = { chains?: Chain[]; options: WalletConnectOptions }
+
+// -- Constants ----------------------------------------------------------------
+const NAMESPACE = 'eip155'
+
+// -- Connector ----------------------------------------------------------------
 export class WalletConnectConnector extends Connector<
-  WalletConnectProvider | UniversalProvider,
+  EthereumProviderType,
   WalletConnectOptions,
   WalletConnectSigner
 > {
@@ -64,16 +51,12 @@ export class WalletConnectConnector extends Connector<
   readonly name = 'WalletConnect'
   readonly ready = true
 
-  #provider?: WalletConnectProvider | UniversalProvider
-  #initUniversalProviderPromise?: Promise<void>
-  #web3Modal?: Web3Modal
+  #provider?: EthereumProviderType
+  #initProviderPromise?: Promise<void>
 
-  constructor(config: { chains?: Chain[]; options: WalletConnectOptions }) {
+  constructor(config: ConnectorConfig) {
     super(config)
-    if (this.version === '2') {
-      this.#createUniversalProvider()
-      if (this.isQrCode) this.#createWeb3Modal()
-    }
+    this.#createProvider()
   }
 
   get isQrCode() {
@@ -81,20 +64,13 @@ export class WalletConnectConnector extends Connector<
   }
 
   get namespacedChains() {
-    return this.chains.map(
-      (chain) => `${defaultV2Config.namespace}:${chain.id}`,
-    )
+    return this.chains.map((chain) => `${NAMESPACE}:${chain.id}`)
   }
 
-  get version() {
-    if ('version' in this.options) return this.options.version
-    return '1'
-  }
-
-  async connect({ chainId }: { chainId?: number } = {}) {
-    const isV1 = this.version === '1'
-    const isV2 = this.version === '2'
-
+  async connect({
+    chainId,
+    pairingTopic,
+  }: { chainId?: number; pairingTopic?: string } = {}) {
     try {
       let targetChainId = chainId
       if (!targetChainId) {
@@ -105,109 +81,34 @@ export class WalletConnectConnector extends Connector<
 
       const provider = await this.getProvider({
         chainId: targetChainId,
-        create: isV1,
       })
       provider.on('accountsChanged', this.onAccountsChanged)
       provider.on('chainChanged', this.onChainChanged)
       provider.on('disconnect', this.onDisconnect)
+      provider.on('session_delete', this.onDisconnect)
+      provider.on('display_uri', this.onDisplayUri)
+      provider.on('connect', this.onConnect)
 
-      if (isV2) {
-        provider.on('session_delete', this.onDisconnect)
-        provider.on('display_uri', this.onDisplayUri)
+      const isChainsAuthorized = await this.#isChainsAuthorized()
 
-        const isChainsAuthorized = await this.#isChainsAuthorized()
+      // If there is an active session, and the chains are not authorized,
+      // disconnect the session.
+      if (provider.session && !isChainsAuthorized) await provider.disconnect()
 
-        // If there is an active session, and the chains are not authorized,
-        // disconnect the session.
-        if ((provider as UniversalProvider).session && !isChainsAuthorized)
-          await provider.disconnect()
-
-        // If there is not an active session, or the chains are not authorized,
-        // attempt to connect.
-        if (
-          !(provider as UniversalProvider).session ||
-          ((provider as UniversalProvider).session && !isChainsAuthorized)
-        ) {
-          await Promise.race([
-            provider.connect({
-              namespaces: {
-                [defaultV2Config.namespace]: {
-                  methods: defaultV2Config.methods,
-                  events: defaultV2Config.events,
-                  chains: this.namespacedChains,
-                  rpcMap: this.chains.reduce(
-                    (rpc, chain) => ({
-                      ...rpc,
-                      [chain.id]: chain.rpcUrls.default.http[0],
-                    }),
-                    {},
-                  ),
-                },
-              },
-            }),
-            ...(this.isQrCode
-              ? [
-                  new Promise<void>((_resolve, reject) =>
-                    provider.on('display_uri', async (uri: string) => {
-                      await this.#web3Modal?.openModal({ uri })
-                      // Modal is closed, reject promise so `catch` block for `connect` is called.
-                      this.#web3Modal?.subscribeModal(({ open }) => {
-                        if (!open) reject(new Error('user rejected'))
-                      })
-                    }),
-                  ),
-                ]
-              : []),
-          ])
-
-          // If execution reaches here, connection was successful and we can close modal.
-          if (this.isQrCode) this.#web3Modal?.closeModal()
-        }
+      // If there is not an active session, or the chains are not authorized,
+      // attempt to connect.
+      if (!provider.session || (provider.session && !isChainsAuthorized)) {
+        await provider.connect({ pairingTopic })
       }
 
-      // Defer message to the next tick to ensure wallet connect data (provided by `.enable()`) is available
-      setTimeout(() => this.emit('message', { type: 'connecting' }), 0)
-
-      const accounts = (await Promise.race([
-        provider.enable(),
-        // When using WalletConnect v1 QR Code Modal, handle user rejection request from wallet
-        ...(isV1 && this.isQrCode
-          ? [
-              new Promise((_res, reject) =>
-                (provider as WalletConnectProvider).connector.on(
-                  'disconnect',
-                  () => reject(new Error('user rejected')),
-                ),
-              ),
-            ]
-          : []),
-      ])) as string[]
+      const accounts = await provider.enable()
       const account = getAddress(accounts[0] as string)
       const id = await this.getChainId()
       const unsupported = this.isChainUnsupported(id)
 
-      // Not all WalletConnect v1 options support programmatic chain switching.
-      // Only enable for wallet options that do.
-      if (isV1) {
-        const walletName =
-          (provider as WalletConnectProvider).connector?.peerMeta?.name ?? ''
-
-        /**
-         * Wallets that support chain switching through WalletConnect
-         * - imToken (token.im)
-         * - MetaMask (metamask.io)
-         * - Omni (omni.app)
-         * - Rainbow (rainbow.me)
-         * - Trust Wallet (trustwallet.com)
-         */
-        const switchChainAllowedRegex =
-          /(imtoken|metamask|omni|rainbow|trust wallet)/i
-        if (switchChainAllowedRegex.test(walletName))
-          this.switchChain = this.#switchChain
-      }
       // In v2, chain switching is allowed programatically
       // as the user approves the chains when approving the pairing
-      else this.switchChain = this.#switchChain
+      this.switchChain = this.#switchChain
 
       return {
         account,
@@ -217,14 +118,7 @@ export class WalletConnectConnector extends Connector<
         ),
       }
     } catch (error) {
-      if (isV2 && this.isQrCode) this.#web3Modal?.closeModal()
-      // WalletConnect v1: "user closed modal"
-      // WalletConnect v2: "user rejected"
-      if (
-        /user closed modal|user rejected/i.test(
-          (error as ProviderRpcError)?.message,
-        )
-      ) {
+      if (/user rejected/i.test((error as ProviderRpcError)?.message)) {
         throw new UserRejectedRequestError(error)
       }
       throw error
@@ -243,25 +137,16 @@ export class WalletConnectConnector extends Connector<
     provider.removeListener('accountsChanged', this.onAccountsChanged)
     provider.removeListener('chainChanged', this.onChainChanged)
     provider.removeListener('disconnect', this.onDisconnect)
-
-    if (this.version === '1' && typeof localStorage !== 'undefined')
-      // Remove local storage session so user can connect again
-      localStorage.removeItem('walletconnect')
-    else {
-      provider.removeListener('session_delete', this.onDisconnect)
-      provider.removeListener('display_uri', this.onDisplayUri)
-    }
+    provider.removeListener('session_delete', this.onDisconnect)
+    provider.removeListener('display_uri', this.onDisplayUri)
+    provider.removeListener('connect', this.onConnect)
   }
 
   async getAccount() {
     const provider = await this.getProvider()
-    let accounts
-    if (this.version === '1')
-      accounts = (provider as WalletConnectProvider).accounts
-    else
-      accounts = (await provider.request({
-        method: 'eth_accounts',
-      })) as string[]
+    const accounts = (await provider.request({
+      method: 'eth_accounts',
+    })) as string[]
 
     // return checksum address
     return getAddress(accounts[0] as string)
@@ -269,8 +154,6 @@ export class WalletConnectConnector extends Connector<
 
   async getChainId() {
     const provider = await this.getProvider()
-    if (this.version === '1')
-      return normalizeChainId((provider as WalletConnectProvider).chainId)
 
     // WalletConnect v2 does not internally manage chainIds anymore, so
     // we need to retrieve it from the client, or request it from the provider
@@ -281,45 +164,12 @@ export class WalletConnectConnector extends Connector<
     )
   }
 
-  async getProvider({
-    chainId,
-    create,
-  }: { chainId?: number; create?: boolean } = {}) {
-    // WalletConnect v2
-    if (this.options.version === '2') {
-      if (!this.#provider) await this.#createUniversalProvider()
+  async getProvider({ chainId }: { chainId?: number } = {}) {
+    if (!this.#provider) await this.#createProvider()
+    if (chainId)
+      this.#provider?.signer.setDefaultChain(`${NAMESPACE}:${chainId}`)
 
-      if (chainId)
-        (this.#provider as UniversalProvider).setDefaultChain(
-          `${defaultV2Config.namespace}:${chainId}`,
-        )
-
-      return this.#provider as UniversalProvider
-    }
-
-    // WalletConnect v1 (Force create new provider)
-    else if (!this.#provider || chainId || create) {
-      const rpc = !this.options?.infuraId
-        ? this.chains.reduce(
-            (rpc, chain) => ({
-              ...rpc,
-              [chain.id]: chain.rpcUrls.default.http[0],
-            }),
-            {},
-          )
-        : {}
-      const WalletConnectProvider = (
-        await import('@walletconnect/ethereum-provider')
-      ).default
-      this.#provider = new WalletConnectProvider({
-        ...this.options,
-        chainId,
-        rpc: { ...rpc, ...this.options?.rpc },
-      })
-      return this.#provider
-    }
-
-    return this.#provider
+    return this.#provider as EthereumProviderType
   }
 
   async getSigner({ chainId }: { chainId?: number } = {}) {
@@ -327,20 +177,12 @@ export class WalletConnectConnector extends Connector<
       this.getProvider({ chainId }),
       this.getAccount(),
     ])
-
-    let provider_ = provider as providers.ExternalProvider
-    if (this.version === '2') {
-      const chainId_ = await this.getChainId()
-      provider_ = {
-        ...provider,
-        async request(args) {
-          return await provider.request(
-            args,
-            `${defaultV2Config.namespace}:${chainId ?? chainId_}`,
-          )
-        },
-      } as providers.ExternalProvider
-    }
+    const provider_ = {
+      ...provider,
+      async request(args) {
+        return await provider.request(args)
+      },
+    } as providers.ExternalProvider
 
     return new providers.Web3Provider(provider_, chainId).getSigner(account)
   }
@@ -357,32 +199,27 @@ export class WalletConnectConnector extends Connector<
     }
   }
 
-  async #createWeb3Modal() {
-    const { Web3Modal } = await import('@web3modal/standalone')
-    const { version } = this.options
-    this.#web3Modal = new Web3Modal({
-      walletConnectVersion: version === '2' ? 2 : 1,
-      projectId: version === '2' ? this.options.projectId : '',
-      standaloneChains: this.namespacedChains,
+  async #initProvider() {
+    const { default: EthereumProvider } = await import(
+      '@walletconnect/ethereum-provider'
+    )
+    this.#provider = await EthereumProvider?.init({
+      showQrModal: this.isQrCode,
+      projectId: this.options.projectId,
+      methods: this.options.methods,
+      events: this.options.events,
+      chains: this.chains.map(({ id }) => id),
+      rpcMap: Object.fromEntries(
+        this.chains.map((chain) => [chain.id, chain.rpcUrls.default.http[0]!]),
+      ),
     })
   }
 
-  async #initUniversalProvider() {
-    const WalletConnectProvider = (
-      await import('@walletconnect/universal-provider')
-    ).default
-    if (typeof WalletConnectProvider?.init === 'function') {
-      this.#provider = await WalletConnectProvider.init(
-        this.options as UniversalProviderOpts,
-      )
+  async #createProvider() {
+    if (!this.#initProviderPromise) {
+      this.#initProviderPromise = this.#initProvider()
     }
-  }
-
-  async #createUniversalProvider() {
-    if (!this.#initUniversalProviderPromise) {
-      this.#initUniversalProviderPromise = this.#initUniversalProvider()
-    }
-    return this.#initUniversalProviderPromise
+    return this.#initProviderPromise
   }
 
   /**
@@ -394,12 +231,7 @@ export class WalletConnectConnector extends Connector<
    */
   async #isChainsAuthorized() {
     const provider = await this.getProvider()
-
-    if (this.version === '1') return true
-
-    const providerChains =
-      (provider as UniversalProvider).namespaces?.[defaultV2Config.namespace]
-        ?.chains || []
+    const providerChains = provider.signer.namespaces?.[NAMESPACE]?.chains || []
     const authorizedChainIds = providerChains.map(
       (chain) => parseInt(chain.split(':')[1] || '') as Chain['id'],
     )
@@ -426,12 +258,9 @@ export class WalletConnectConnector extends Connector<
           }),
         ),
       ])
-      if (this.version === '2') {
-        ;(provider as UniversalProvider).setDefaultChain(
-          `${defaultV2Config.namespace}:${chainId}`,
-        )
-        this.onChainChanged(chainId)
-      }
+      provider.signer.setDefaultChain(`${NAMESPACE}:${chainId}`)
+      this.onChainChanged(chainId)
+
       return (
         this.chains.find((x) => x.id === chainId) ?? {
           id: chainId,
@@ -467,5 +296,9 @@ export class WalletConnectConnector extends Connector<
 
   protected onDisplayUri = (uri: string) => {
     this.emit('message', { type: 'display_uri', data: uri })
+  }
+
+  protected onConnect = () => {
+    this.emit('connect', { provider: this.#provider })
   }
 }
