+++ packages/connectors/tsup.config.ts
@@ -15,6 +15,7 @@ export default defineConfig(
       'src/mock/index.ts',
       'src/safe.ts',
       'src/walletConnect.ts',
+      'src/walletConnectV1.ts',
     ],
     external: [...Object.keys(dependencies), ...Object.keys(peerDependencies)],
     platform: 'browser',
